+++ spec/rails_app/db/migrate/core/20101224223620_create_users.rb
@@ -6,7 +6,7 @@ class CreateUsers < ActiveRecord::Migration
       t.string :crypted_password
       t.string :salt
 
-      t.timestamps
+      t.timestamps                :null => false
     end
   end